Changed field pool->serviceclass name.
This is the service class to where the file have to be written to. It is not
possible to specify directly the target tapes or tapepools, but we can use
the information stored in the service classes.

By default, the files are written to the tapepool belonging to the repack
service class specified in the castor config file.