- fixed memory cleaning of Repack Objects - fixed bug: non init of stager port in stager_options caused error in communication - we log now in DLF if files in invalid status were found
still there is a problem with the stager_qry call. It fails sometimes due to a dynamic cast in the c++ part and the returned number of responses is not correct. This causes then a crash, because the function to free tries to delete the amount of objects given by this number.
=> in this case the answer is not deleted !! <=