Added internalputDoneFunc: in fact an extraction from putDoneFunc. It creates a specified number of tapecopies for a diskcopy in STAGEOUT. This number is passed as a parameter.
Changed fileRecalled:
Whenever a file is recalled its repackvid in the subrequest is checked and
for each found 'repack-subrequest' for this file the number of tapecopies
are created. Only subrequest in WAITTAPERECALL,WAITSUBREQ and READY are
considered. This is neccessary for >1 tapecopy repacking (otherwise only
one tapecopy is created for the migration).


Changed putDoneFunc:
Reduced by the extracted part now in internalputDoneFunc