Implementation for checksumming on the fly for rfiod: For PUT, we are calculating  it  in  the consume64_thread function  after writing  the file locally  to the disk, and update extended  file attributes with a user.castor.checksum.value and a user.castor.checksum.type.
For GET, we are also calculating checksum on the fly in the produce64_thread function after reading the file locally from the disk. In the end of reading(transferring) we compare this checksum with the checksum from the file attribute, if they are not equal we set errno=EREMOTEIO (Remote I/O error) and rfcp fails with return code 2(system error) and message Remote I/O error.

To use checksumming in rfiod we have to set RFIOD USE_CKSUM YES (castor.conf) and  should have file  system  that allow to set extended attributes (for the filesystem without xattr support, we will ignore checksumming).