The method setAndSendIoResponse receives an extra parameter std::string physicalFileName (that modification is required by xrootd)
In the reply we are always sending an IOResponse when it should be a FileResponse or IoResponse depending on the type of the request.
In fact, this method will make more sense in the Response object, and depending on the Request the handler fills the specific Response and send it from there.
Till now there is not problem as in the setAndSendIoResponse method only the fields/attributes specific of FileResponse are field, so when you cast on the client side you don't lose information. In the case of xrootd in the client api they cast into a IOResponse, so that should be taken into account if one day this method is rewritten.