My last check-in of CASTOR2/castor/Constants.hpp could be termed as a "nice try" or "close but no cigar".  Yes I broke the build... Sorry!
One can use a constant as the size of a variable outside of a function with
C++, as the following code which compiles successfully with -pedantic -ansi
demonstrates:

static const unsigned int ObjectsIdsNb = 152;

const char* a[ObjectsIdsNb];

int main(int argc, char **argv) {
  using namespace std;

  cout << "Hello World!" << endl;
  cout << "sizeof(a)/sizeof(char*)=" << sizeof(a)/sizeof(char*) << endl;
}

Unfortunately the C compiler, as opposed to the C++ compiler, does not allow a
variable-size type declared outside of any function.

In order to keep a solution where the numbers of object, service and
repriesentation IDs are only written once by the programmer, I have added a
macro for each of these values whilst preserving the C++ constants.  For
example, in the case of ObjectsIdsNb the code now reads:

  /**
   * Nb of objectIds currently existing
   *
   * Note that both a macro and a constant are required because the C compiler,
   * as opposed to the C++ compiler, does not allow a variable-size type
   * declared outside of any function.
   */
#define OBJECT_IDS_NB 152
  static const unsigned int ObjectsIdsNb = OBJECT_IDS_NB;