Modified the showqueues command to display negative (-1) delta times as zero.
The showqueues command calculates delta times by calculating the difference
between absolute times given by the VDQM server and the current time
determined by the showqueues command on the client machine.  The VDQM2 is
rather fast at returning the results required by showqueues.  As a result,
delta times of -1 are displayed from time to time due to a clock skew of
less than a second between the client and server machines.