Added -c and -h options to rtcpclientd.
The -h option just displays the usage message without giving an unknown
option error.

The -c option allows the user to overide the default location of the castor
configuration file.  This is useful when one wants to run a locally compiled
version of a daemon on a host with an rpm installation of the same daemon.  One
needs to make a castor configuration file with the locations of the locally
built dynamic shared libraries and then use the -c option to point to it.