Savannah bug bug #43056: instead of relying on /proc/scsi/scsi in order to find a matching pair of st and sg device nodes, rely on sysfs.
Except for right after boot, the order of devices in /proc/scsi/scsi is not
guaranteed to be the same as in /dev/sgX. So, it can easily happen that the
code sends SCSI commands to a disk (which already happened) or to the wrong
tape drive.

By this change the support of the CASTOR tape s/w is limited to Linux kernels
2.6 and higher. Code for other architectures and operating systems has been
removed.