A class's destructor is always called before the destructors of it non-static members:
  http://msdn.microsoft.com/en-us/library/8183zf3x(VS.80).aspx

This being so, I have removed the dynamic memory management of the queue
within the DynamicThreadPool.  The destructor of the DynamicThreadPool waits
for the threads popping tasks from the queue to complete before continuing to
its end.  Only aifter the destructor has finished does the compiler code then
call the destructor of the queue, therefore removing the need to explicitly
manage the memory allocation of the queue using new and delete.