The tape aggregator now has two port number constants; one for listening to connections from the VDQM and one for listening to connections from RTCPD.
Two ports are required because of the major differences between the VDQM and
RTCOPY prototcols.  In the case of the VDQM protocol, the tape aggregator
should listen, accept and then read in an incomming VDQM message, in other
words the magic number, request type, message body length and finally the
message body.  In the case of the RTCOPY protocol, the tape aggregator should
listen, accept and the send as opposed to read in a message.