I have moved the code in RtcpdHandlerThread to VdqmRequestHandlerThread and have removed RtcpdHandlerThread.  The code compiles, but probably does not work.
The concurrency and communication model has been decided.  There will be a pool of VDQM threads with one thread per tape drive.  Today there is only one drive
per tape server at CERN, but we are trying to keep backwards compatibility.
The existing code we are trying to replace kept open the option of more than
one drive per tape server.  One VDQM thread will use a select loop to
communicate with the five or more incomming connections required by RTCPD.  By
default, RTCPD makes five connections to a client.  The initial connection to
signal the start of a remote tape copy and to listen for ping and abort
messages, one connection for the RTCPD tape IO thread, and three connections
for the default number of three RTCPD disk IO threads.

The next step is to code the select loop for the five or more connections.