Added another macro, this time called CASTOR_DLF_WRITEC, notice the lack of the letter P at the end of the name.  There are now two macros to help generate log messages with the contextual information of file, line and function:
CASTOR_DLF_WRITEC
CASTOR_DLF_WRITECP

The first macro logs a message without parameters accept for the three
contextual ones of file, line and function.  The second macro logs a message
with parameters to which the three contextual ones are prepended.

The meaning of the letter C at the end of the CASTOR_DLF_WRITEC is 'C'onextual
log.  The meaning iof the letter P at the end of CASTOR_DLF_WRITECP is with
'P'arameters.