Renamed and improved the documentation of the following function:
  processRtcpFileReq

This function basically implements the common logic of the rtcpFileReqCallback
and rtcpFileErrReqCallback functions.

Corrected a bug in the new processRtcpFileReq function.  When sending
acknowledge messages to RTCPD it has to use the value of the request type
passed into it via the "header" parameter.  This is because the
processRtcpFileReq function has to handle both RTCP_FILE_REQ and
RTCP_FILEERR_REQ messages which have different request type values.  The
request type of the "header" parameter is set to appropriate value before the
processRtcpFileReq function is called.