Savannah bug #32083: added code to make use of the MTIOCSENSE ioctl to retrieve the sense bytes from the SCSI layer. This may be helpful in order to tune reactions in certain error scenarios, e.g. set a tape read-only if there are medium errors.
The kernel part of the ioctl is provided by a patched st driver
(kmod-st-driver, provided by Linux support). If this patched driver is
not available (i.e. not installed), the code will fall back to the old
ioctl calls.

In order to leave the standard include files untouched, the ioctl's
definition is in an added file (h/mtio_add.h). Once the ioctl is part of
the official kernel, this include is not needed anymore and can be removed.
It is only included if the standard includes do not have a definition
for MTIOCSENSE.