This commit fixes the following bug:
bug #54485: rtcpd dumps in error handling on SLC5

I modified the rtcpd_FreeBuffers() function so that it does not free the locks
used to protect the rtcpd data buffers and to broadcast exceptions to all of
the rtcpd threads.  The reason for this modification is the CLThread
concurrently broadcasts an exception using the locks whilst the tape IO thread
calls rtcpd_FreeBuffers() which deletes the memory used by the locks and sets
the intermediate pointers to the locks to NULL.  Many core dumps have resulted
from these concurrent and conflicting actions.