This commit addresses the imporovement of the vmgrdeletetape security checks for the following RFE:
bug #58794: RFE: Make vmgrdeletetape name server safe

It has been created a new request type for the vmgrdeletetape in order to force the use of the new vmgrdeletetape version and report an error of "incompatible versions" otherwise (when the server run the new VMGR version).
In the case the user run the new vmgrdeletetape version versus an old VMGR server version, the error will be: "VMG02 - illegal function 0".