This modification is related to the Support Request:
sr #112306: locking issue in CMS stager

The restartStuckRecalls procedure is now been moved fron the file oracleGC.sql 
to the file oracleTape.sql and has comments explaining the role of this procedure.
The code for the creation of the database job RESTARTSTUCKRECALLSJOB has been 
also moved in the file oracleTape.sql.

The actual SQL code of the restartStuckRecalls procedure has been modified to
now cover more cases of stuck recall requests.