This commit works towards fixing the following bug:
    bug #50008: Incorrect logging of nsfileid and nshost

The TapeBridgeDaemon has been refactorised to help test the Dlf client API.

The tapebridge::LogHelper continues to log nshost and fileid as extra DLF
parameters because it is responsible for logging the contents of legacy
messages preserving the contained fields and their order.  To help towards
fixing bug bug #50008, the tapebridge::LogHelper now passes the fileid and
nshost as parameters to the appropriate dlf_writep function so that the DLF
system can efficiently catalogue the logs in question via fileid and nshost.