This commit fixes the following bug:
bug #65928: Deadlock between rtcpclientd ending a stream and mighunterd adding work to a stream

The bug is fixed by catching the deadlock detected exception and treating it
in the same way as not being able to delete a stream because it has tape copies.
This bug fix works because the stream thread of the mighunter daemon
periodically cleans up stopped streams with no tapecopies.