This commit works towards fixing the following bug:
bug #65928: Deadlock between rtcpclientd ending a stream and mighunterd adding work to a stream

The tape gateway was not taking a lock on a stream before trying to delete it.
This lock is necessary because the tape gateway uses a delete from stream
statement which can grab multiple implicit locks on the Stream2TapeCopy table
if there are attached tape copies at the time of the delete.