This commit completes the work to address the following BUG:
bug #67740: The SQL view used by showqueues returns duplicate rows for multi-dedications

The bug is caused by the fact that the TapeDriveShowqueues_VIEW still includes a join with the TapeDriveDedication table which has one row per dedication. The join is no longer used as the castorVdqmView.getVdqmDedicate(TapeDrive.Id) function replaces it.