The StreamThread now gracefully shuts down the migration hunter daemon in the following 3 cases:
* The streamPolicy attribute of the service class is an empty string
* The streamPolicy function is not in Python-module
* The stream policy could not be applied

I have also fixed a bug in the MigHunterThread where an incorrect test for
NULL was made when dealing with an exception thrown by a failing execution of
the migration policy.  There was simply no need to make the test.

This commit works towards implementing the following RFE:

bug #68020: RFE: The mighunter should gracefully stop when it encounters an invalid migration-policy or stream-policy function-name