bug #59706: Error in dumptp, blockID retrieval when blockid>0 generates an error.
Corrected the handling of the sg device file name which got lost as it was extracted from a structure that got overwritten to pass out the result. The file name is now explicitely extracted beforehand and stored in a local variable. Changed the signature of the rtcpd_DmpFile function which is only called from a single place.