bug #66570: RFE: need number of running streams per tape pool in policy
Added the svcClassNbDrives parameter to the stream-policy Python-function.  This
parameter is required due to how streams are created within the CASTOR stager
database.  Unfortunately the effective total-number of streams created for a
given service-class may well exceed the nbDrives attribute of the service class.
This feature is due to the fact that multiple service-classes can share the
same tape-pool and extra streams can be created due to race conditions and by the
simple fact that one service-class may wish to create N streams for a shared
tape-pool, whereas another may wish to create N+M, and the later will be the
dominant in this case.  The svcClassNbDrives attribute should be used to try and
cap the number of running streams, though of course the stream with the biggest
value for nbDrives will have the final say.