bug #69634: RFE: Add TPVID to as many tape-bridge logs as possible
Compilation of the  tape-bridge works on our development and
TeamCity build machines, but randomly does not on other SLC5
boxes.  The cause of this randomness is some of the code
compiled into the tape-bridge, tape-gateway and tape-bridge
clients should be done once in a library as opposed to being
compiled 3 different times via different makefiles.  This
commit finishes the above RFE by making the build process
more stable by putting the files in question in libraries.