bug #66570: RFE: need number of running streams per tape pool in policy
Tests have shown that there is a very "bad feature" / bug when the
stream-policy logic enters "emergency mode" and that this feature existed
before this RFE.

The bad feature is explained by the following code snippet from release 2.1.9-8
and more specifically SVN file castor/tags/v2_1_9_8/castor/db/oracleTaple.sql:

    -- emergency mode
    OPEN dbInfo FOR
      SELECT Stream.id, 10000, 10000, gettime
        FROM Stream
       WHERE Stream.id IN
         (SELECT /*+ CARDINALITY(stridTable 5) */ *
            FROM TABLE(strIds) stridTable)
         AND Stream.STATUS = 7
       GROUP BY Stream.id;

Emergency mode is entered when there is too big a back-log of tape
migration-jobs that calculating statitics such as the total number of bytes to
be written to tape or the age of the oldest migration-job is simply too
costly.  In this case the above logic assumes 10000 migration-jobs with an
average file-size of 1 byte and an age equal to the number of second since the
unix epoch (why?).

The following code snippte from this commit shows that this commit gives
"emergency mode" some better assumptions, 10000 migration-jobs as before, but
this time with 1 GiB as the average file-size and an age of 48 hours for the
oldest job.

    -- Enter emergency mode
    OPEN outStreamsForPolicy FOR
      SELECT Stream.id,
             10000, -- numTapeCopies
             10000*1073741824, -- totalBytes (Force file size to be 1 GiB)
             48*3600, -- ageOfOldestTapeCopy (Force age to be 48 hours)
             Stream.tapepool
        FROM Stream
       WHERE Stream.id IN (
                SELECT /*+ CARDINALITY(streamIdTable 5) */ *
                  FROM TABLE(varStreamIds) streamIdTable)
         AND Stream.status = 7;  -- STREAM_WAITPOLICY

This commit also corrects a comment in the following file concerning the age of
the oldest tape migration-job:

    /etc/castor/policies/streams.py.example