bug #66570: RFE: need number of running streams per tape pool in policy
I am putting back the original logic used for selecting candidates for the
stream-policy.

The original logic selected streams which were either newly-create, waiting for
space or stopped.

The new logic I introduced with this RFE only selected streams waiting for
space or stopped.  The choice to not select newly created streams was to fix a
temporal bug that I observed during testing.

With the original logic, a backlog of more than 10,000 files, no streams yet
created and a period of 1 second for both the MigHunterThread and the
StreamThread, the following temporal bug would occur.  The MigHunterThread
would see the large number of files to be migrated, it would then create a set
of new streams, and then start to call the migration-policy python-function for
each of the files to be migrated.  This calling of the policy would take longer
than the stream thread doing its work.  The StreamThread would select the
newly created streams with no tape-copies attached yet (the MigHunterThread
would be  still determining them by making a relatively expensive statistics
query on the database followed by calling the migration-policy python-function
many thousands of times).  Given a newly created stream with no tape-copies
the StreamThread would then happily delete it therefore preventing the
MigHunterThread from ever attaching any tape-copies to it.  The net result is
a mighunter which goes nowhere very fast.

The new logic attempted to fix the problem by having the StreamThread leave
newly created streams alone.  Unfortunately I discovered that it is not
guaranteed that all of the tape-copies to be attached to a service-class'
streams are in fact attached to all of the streams.  My apparent fix therefore
unwittingly introduced a leak of newly created streams with no tape-copies
attached to them.

I am putting back the old logic knowing that I will put back the old temporal
bug as the old bug will happen rarely in production and giving a relatively
short period to the MigHunterThread and relatively long period to the
StreamThread will mean the bug hardly ever happens.