bug #54597: RFE: ONHOLD explanation option
I broke the build on my last commit, by adding the value GET_MULTICOPYINFO to
the framework generated enumeration castor::repack::RepackCommandCode.  The
following method of the RepackClient contains a switch statement on the
enumeration which does not have a default clause.  The compiler correctly and
very nicley detects the error that not all of the enumeration's value are
accounted for.

void RepackClient::handleResponse(castor::repack::RepackAck* ack)

I have now added two placeholders in the switch statement.  One for
GET_MULTICOPYINFO and one for default.  These placeholders need to fleshed out
before bug #54597 can be said to be fixed.