bug #75722: RFE: The VMGR should support tape capacities greater than 2TB
Added test of adding a 5T tape.  This test includes verifying the output of
vmgrlisttape and vmgrlistpool for both old and new versions of the command-line
clients.  The test take into account the fact that the old version of
vmgrlisttape will not see a tape size larger than 2T as the new vmgrd daemon
sets a ceiling of 2T for old vmgrlisttape clients as the underlying TCP/IP
protocol does not support larger sizes.