bug #75722: RFE: The VMGR should support tape capacities greater than 2TB
Corrected the vmgr upgrade script.  When calculating the total amount of free
space in a tape-pool one must only sum the free space of tapes which are either
FREE or BUSY.  Before this modification the upgrade script was sum the free
space of all tapes within a pool which was incorrect.