bug #80262: [RFE] Too much logging with above-DEBUG level in tapegateway.
The downgrading of the log messages was too drastic, hence the information
about the requester, which was logged separately, got lost. To fix this problem
we now pass down this information to the handlers, which will log it systematically.
In the mean time, the handler map mechanism has been removed to replace it with
a simpler switch statement, which allows easier debugging.



Fixed a typo in a comment.

Trying eclipse & teamcity's autocommit on successful build.