bug #82673: RFE: Tapebridged and rtcpd should support buffered tape-marks over multiple files
Improved the error message generated when rtcpd un-expectedly closes the initial
connection to tapebridged.  Before this fix tapebridged was incorrectly
reporting this error as un-exepected data received from the initial connection.
The reason behind this bug was a socket descriptor passed to the select()
function can become pending because its associated connection been closed as
well as data being available for reading.