bug #82673: RFE: Tapebridged and rtcpd should support buffered tape-marks over multiple files
To help implement this RFE I have moved the function named
rtcpd_SignalFilePositioned into its own couple of source and header files.  This
move has also removed the function's dependency on the global variable named
proc_cntl.  A pointer to this variable is now cleanly passed into the function
as a parameter.