bug #84322: Lack of per-file error reporting system in tapegtewayd/tapebridged
The cprc (file copy return code) of the rtcpd message called RtcpFileRqstMsg will needs to be interpreted by the tapebridged in order to implement a fix for this bug.  It's datatype is currently wrong, it is defined as being uint32_t when it should be signed.  The following commit fixes this problem on the trunk.