bug #82673: RFE: Tapebridged and rtcpd should support buffered tape-marks over multiple files
This commit modifies the tapebridged/rtcpd protocol that is internal to a
tape-server.  The new version of the protocol is now capable of supporting
buffered tape-marks, though the version of rtcpd in this commit still only
supports at best one synchronised tape-mark per file, i.e. it does not support
buffered-tape marks over more than one file.

The change with respect to the previous version of to the tapebridged/rtcpd
protocol is simple.  When the rtcpd daemon knows that its client is is the
tapebridged daemon it now sends an explicit TAPEBRIDGE_FLUSHEDTOTAPE message to
the tapebridged daemon when it can guarantee one or more files have been
flushed to tape.  The tapebridge daemon will now only send
FileMigratedNotification messages to the tapegatewayd daemon for the files that
have been explicitly flushed to tape.  The PendingMigrationsStore is used to
hold back FileMigratedNotification messages until the appropriate
TAPEBRIDGE_FLUSHEDTOTAPE message has been received.  The rtcpd daemon in this
commit always does at least one synchronised tape-mark per file and sends
always sends one TAPEBRIDGE_FLUSHEDTOTAPE per file when its client is the
tapegatewayd daemon.