bug #85285: tapegateway::WorkerThread::handleMigrationUpdate blindly sends a positive acknowledgements
Fixed the failing VMGR update when the tape was disabled. Reduced the cases when the vmgr helper functions refuse to work.
The vmgr helper class of the tape gateway was heavily refactored in the process.
Also added a mechanism to prevent looping on vmgr in case of a shutdown of the demon.
This was implemented with a new functor in tape/utils.
Homogenized the authorship information in tape gateway directory (to castor-dev team).