bug #75494: RFE: Archiving a tape should cancel all outstanding recall requests for that tape
The cancelRecallForTape procedure has been extended to also change the status of the tape involved in the
DB. This will prevent rtcpclientd from looping on the vmgr.

The procedure cancelRecall has been made safer in the process as in some (unusual, but possible) cases, it would have killed migrations too, if any.