bug #85949: RFE: Add bulk messages to tapegatewayd/tapebridged protocol
Added new bulk messages to RH.xmi and auto-generated the associated classes.  Please note that there is no graphical information in RH.xmi concerning the new messages because gencastor went mad and changed the order of the members of existing classes.  We now add new content manually for readablility, sanity and maintainability.