bug #86163: RFE: Tape-server software should support both Spectra logic and virtual tape libraries
Small fixes for CDB commands to work correctly with Spectra library and  mhvtl.
As soon as a Spectra library and Virtual Tape Library are close loop libraries it was added special vriable to castor.conf to switch to this logic.
Right now it is just a fix for human presentation of loaded drives.

Also was fixed incorect error handling for 0xB6 CDB command.