bug #86163: RFE: Tape-server software should support both Spectra logic and virtual tape libraries
We will skip the 0xB6 command (SEND VOLUME TAG) for Spectra like library as soon as it is not supported.
This makes smc_find_cartridge function work much faster.
A variable "SMC LIBRARY_TYPE SPECTRA" in castor.conf will switch the logic.

The Virtual tape library also does not support 0xB6 so this fix is good for it too.