bug #90313: RFE: tapebridged should request more files to transfer in bulk
The implementation of this RFE is going to take a relatively long period of
time in that there will be several commits to SVN before it is completed.

This commit adds some helper classes to the tapebridged source-code base.
These classes are tested by the unit-tests of the test/unittest directory but
are not actively used by the tapebridged daemon and so will not effect any
release that may need to be made of the trunk.