bug #90313: RFE: tapebridged should request more files to transfer in bulk
Fixed a stack-overflow bug in the CASTOR 2.1.12-2 internal release of the
tapebridged daemon.  The bug was caused by an unintentionally recursive
function.  The bug was not detected on my development system but was detected
on castorcert5 because the bug only manifested itself when an error was
reported to the tapebridged daemon by the rtcpd daemon.