bug #93680: NsTapeGatewayHelper::checkRecalledFile causes a segmentation fault when file has no checksum in the name-server
After a discussion with other CASTOR developers it has been decided to
temporarily modify the tapegatewayd daemon so that it allows the recall of an
old tape-segment with no name-server checksum to complete succesfully and
without any attempt by the tapegatewayd daemon to record the newly calculated
checksum in the name server.

In the next major release of CASTOR, the tape-segment recall logic will be
re-written and this new logic will record the newly calculated checksum in the
name server.