bug #92460: tapebridged should gracefully shutdown a migration tape-session when tapegatewayd reports a disabled tape
Refactored the tapebridged daemon and created the following unit-test that
recreates this bug:

castor::tape::tapebridge::BridgeProtocolEngineTest::testMigrationToDisabledTapeUsingLocalDomain

The refactoring was necessary in order to create the unit-test.  During the
refactoring I modified the logging of the tapebridged daemon so that the very
first session error detected is always logged immediately.  This will help
debug problems in the future if need be.