Rewrite the db interface and worker thread of tape gateway to implement the bulk update inside the DB. Interim check-in. Bulk is still a work in progress.
DB interface relies on setVector in OCCI, which requires the OBJECT option to be enabled in the client.

Also:
Fixed remaining class declaration for Tape.
Fixed modifications of messages in DLF which.
Expanded the occiNumber utility class to bring full 64 bits support without the limitations of the cast to a double (was 51 bits).