bug #92460: tapebridged should gracefully shutdown a migration tape-session when tapegatewayd reports a disabled tape
Tests for the 2.1.13-0 release have shown that the tapebridged daemon does not
cope correctly with a tape being DISABLED just before it asks for the first
file to be migrated.  The tapebridged daemon incorreectly ends the tape session
with an EndNotification message when it should end the tape session with a
EndNotificationErrorReport.

This commit adds the unit-test that fails if it detects this bad behaviour.