bug #96536: rmlabelinfo() of tape/initlabel.c corrupts memory if called more than once
rmlabelinfo() assumes we support multi-drive reservations and contains buggy
memory managment to deal with these cases.  Such reservations are no longer used
therefore the buggy code can be removed.

This commit adds a unit test which fails (test driven development) because the
future implementation of initlabelroutines() should only accept 1 as a parameter.
initlabelroutines() is the creator of the C structures that rmlabelinfo() tries
to remove.