bug #96536: rmlabelinfo() of tape/initlabel.c corrupts memory if called more than once
The BridgeProtocolEngine::rtcpTapeErrReqRtcpdCallback() method currently
aborts the tape session if it detects an error by throwing an exception.
This is too abrupt and should be changed to trying to gracefully shutdown
the rtcpd session.

This commit introdcues the graceful shutdown of the rtcpd session in the
event of an error detected by the
BridgeProtocolEngine::rtcpTapeErrReqRtcpdCallback() method.