bug #96742: RFE: rtcpd should be configurable as to whether or not it tries to reuse tape mounts
The tapeNeedsToBeReleasedAtEndOfSession variable of the rtcpd_MainCntl()
function needs to be a static because its address is passed from the main
thread to the tape IO thread.  Making the variable static avoids the
possibility of memory corruption if the rtcpd_Main() function ends
prematurely before the main thread calling it has joined with the tape IO
thread.