bug #88114: RFE: wrong compression for LTO5 and Oracle T10KC drives
I have improved the tapebridged side of the batch compresion logic by adding
code to deal with the unlikely case that the sum of the original file sizes of
a batch of migrated files is zero bytes.  This should never happen, however the
code should not crash if presented with such a case.

The following two new unit tests explain the expected behaviour and pass with
success.

  CPPUNIT_TEST(testCalcMigrationCompressionRatioWithZeroFileSize);
  CPPUNIT_TEST(testCorrectMigrationCompressionStatisticsWithZeroFileSize);