bug #99219: RFE: A file should never be reported as having a compressed file size less than 1 byte
Following the test-driven development methodology this commit adds the following
failing unit-test:

    testSetMigrationCompressedFileSizeWith1B1MBAnd0Point4Ratio

This test passes two files, one of 1 byte and one of 1 megabytes into the
following method along with a number of bytes written to tape equal to a
quarter of a megabyte.

BridgeProtocolEngine::setMigrationCompressedFileSize()

The compression ratio will be less than 0.5.  This means that without a
floor of 1 byte the compressed size of the 1 byte file would be zero
bytes, because 1 byte times a value less than 0.5 will round down to zero
bytes.  The aforementioned method must implement a floor of 1 byte and
therefore return 1 byte and not zero bytes.