bug #99787: RFE: Add ProcessingTime to gateway log "Worker: file to recall retrieved from db"
The per-file log cannot be assigned a processing time, as it is a printout of the individual content from a bulk list. The global processing time, plus the per-file average were already available, but in DEBUG level.

The summary for recalls also wrongfully used the migration summary message.

2 messages, one of them new, have hence been promoted to Info level:

Worker: files to migrate retrieved in bulk from db
Worker: files to recall retrieved in bulk from db

They come AFTER the listing of the individual files.