Bug #99889: RFE: Enable support for read-only hardware
- Introduced support for a READONLY state for Diskservers and Filesystems.
Hardware in such a state will accept migrations, [PrepareTo]Get and
DiskCopyReplica requests, and will not accept recalls or [PrepareTo]Put/Update
ones when they switch to write mode.
- Dropped the {DiskServer,FileSystem}.adminStatus fields and the obsoleted
code around them and introduced a new DiskServer.hwOnline boolean field, which
is updated by the heartbeat reports and displayed by printdiskserver.
In turn, modifydiskserver can not change the online flag but only the status.
For backward compatibility, on stager_qry -s a Diskserver will be reported
as DISABLED if hwOnline = 0 regardless its status.