bug #101536: The tapebridged daemon does not handle zero length files for migration correctly.
The new unit tests (there are now only 3) for the PendingMigrationsStore class
have been improved and the PendingMigrationsStore class itself has been
sucessfully modified to pass the tests.

The next step is propagate the new SessionException up through the call stack
until the BridgeProtocolEngine::runRtcpdSession() method and modify the method
so that it extracts the SessionError wrapped by the SessionException and
appends it to BridgeProtocolEngine::m_sessionErrors memebr variable.