Updated the ReleaseNotes.
Please note that I have removed the following RFE:

    #98275  RFE: rtcpd should support an xroot protocol for data transfers

This is because the following RFE has reversed its changes:

  bug #101663: Temporarily remove tape support for the xrootd protocol

In order not to confuse the users, neither RFE should be mentioned in the
ReleaseNotes as they have effectivley annihilated each other.

The current tape server daemons of CASTOR will probably never talk xroot.