Changed the Drive classes layout:
Tape::Drive is now a container, taking care of allocating the proper type of drive depending on the device information.
Tape::DriveGeneric is the old parent class. It is now a pure virtual for lack of the non-common functions.

The tests (unit and system) have been adapated.

The system test environement now has a vide sampler of drive types (LTO from IBM and HP, 3592 from IBM in 3 flavors and T10K A, B and C from Oracle). The complete list of flavors is in usr/vtltape.c in mhvtl sources.